+++ src/deepsparse/v2/text_generation/__init__.py
@@ -21,6 +21,7 @@ from .join_output import *
 from .kv_cache_operator import *
 from .multi_engine_prefill_operator import *
 from .nl_engine_operator import *
+from .nl_engine_operator_no_kv_cache import *
 from .prep_for_prefill import *
 from .process_inputs import *
 from .process_outputs import *
@@ -30,3 +31,4 @@ from .token_generator import *  # isort:skip
 from .prep_for_generation import *  # isort:skip
 
 from .pipeline import *  # isort:skip
+from .pipeline_no_kv_cache import *  # isort:skip
